fix: fixed require cycle problem
