+++ src/contexts/financesContext.tsx
@@ -2,14 +2,13 @@ import React, { createContext } from "react";
 import Loading from "../components/global/loading";
 import { FinancesContextData } from "../interfaces/screens/finance.interface";
 import { FinanceBalance } from "../interfaces/services/finance.interface";
+import { AppBalanceService } from "../services/balance";
 import { AppFinanceService } from "../services/finance";
-import { DebtsContext } from "./debtsContext";
 
 export const FinancesContext = createContext<FinancesContextData>({} as FinancesContextData);
 
 export const FinancesProvider = ({ children }: any) => {
 
-    const { getDebtsBalance } = React.useContext(DebtsContext)
     const [filteredMonth, setFilteredMonth] = React.useState<string>('')
     const [filteredYear, setFilteredYear] = React.useState<string>('')
     const [finances, setFinances] = React.useState<FinanceBalance>();
@@ -22,9 +21,8 @@ export const FinancesProvider = ({ children }: any) => {
     const getFinancesBalance = async () => {
         try {
             setLoading(true)
-            const financesBalance = await AppFinanceService.getFinancesBalance(filteredMonth, filteredYear, 1);
+            const financesBalance = await AppBalanceService.getFinancesBalance(filteredMonth, filteredYear, 1);
             setFinances(financesBalance)
-            getDebtsBalance()
             return financesBalance;
         } catch (error) { } finally {
             setLoading(false)
