+++ src/services/debts.ts
@@ -1,12 +1,9 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { DEBTS_INSTITUTION } from "../constants/debts.constants";
 import { ASYNC_DEBTS } from "../constants/storage.constant";
-import { DebtsBalance, DebtsDto, DebtsEntity, DebtsInstitution } from "../interfaces/services/debts.interface";
-import { FinancesBalanceEntity } from "../interfaces/services/finance.interface";
+import { DebtsDto, DebtsEntity, DebtsInstitution } from "../interfaces/services/debts.interface";
 import { Services } from "../interfaces/services/service.interface";
 import { getPipeDateTimeString } from "../utils/date.util";
-import { AppCategoryService } from "./category";
-import { AppFinanceService } from "./finance";
 
 class Debts implements DebtsEntity {
     id: number;
@@ -73,51 +70,6 @@ class DebtsService implements Services<DebtsEntity, DebtsDto> {
         await AsyncStorage.setItem(ASYNC_DEBTS, JSON.stringify(remove));
 
         return remove;
-    }
-
-    public async getDebtsBalance(): Promise<DebtsBalance[]> {
-        const finances = await AppFinanceService.find();
-        const categories = await AppCategoryService.find();
-
-        const debts = await this.find();
-        const debtsBalance: DebtsBalance[] = []
-
-        for (const debt of debts) {
-            if (!debt.totalPerMonth) debt.totalPerMonth = debt.total
-            const institution = DEBTS_INSTITUTION.find(institution => institution.id == debt.institutionId) as DebtsInstitution;
-
-            const financesFilter = finances.map(finance => {
-                if (finance.billId == debt.id) {
-                    const category = categories.find(category => category.id == finance.categoryId);
-                    const bill = DEBTS_INSTITUTION.find(institution => institution.id == debt.institutionId)
-                    return { ...finance, category, bill }
-                }
-            }).filter(r => r) as FinancesBalanceEntity[];
-
-            let totalPaid = 0;
-            financesFilter.map(finance => { if (finance.type == "EXPENSE") totalPaid += finance.value })
-
-            const getTotalMonth = () => {
-                if ((debt.total - totalPaid) > 0) return Math.ceil((debt.total - totalPaid) / debt.totalPerMonth)
-                return 0
-            }
-
-            let totalRemain = 0;
-            if ((debt.total - totalPaid) > 0) totalRemain = debt.total - totalPaid;
-
-            const totalMonth = getTotalMonth()
-
-            debtsBalance.push({
-                ...debt,
-                totalMonth,
-                institution,
-                finances: financesFilter,
-                totalPaid,
-                totalRemain
-            })
-        }
-
-        return debtsBalance
     }
 
     protected findLast(debts: DebtsEntity[]): DebtsEntity | undefined {
